+++ lib/src/model_viewer_plus_web.dart
@@ -5,10 +5,8 @@ import 'package:flutter/services.dart' show rootBundle;
 
 import 'html_builder.dart';
 
-// import 'dart:ui' as ui;
 import 'shim/dart_ui_fake.dart' if (dart.library.html) 'dart:ui' as ui;
 import 'shim/dart_html_fake.dart' if (dart.library.html) 'dart:html';
-// import 'dart:html';
 
 import 'model_viewer_plus.dart';
 
@@ -39,7 +37,6 @@ class ModelViewerState extends State<ModelViewer> {
             // Loading Attributes
             'src',
             'alt',
-            'poster',
             'poster',
             'seamless-poster',
             'loading',